Rename snake-case filenames to camelCase filenames
