+++ src/components/itemCompactView.js
