+++ src/components/itemDetailedView.js
